Fix developer crate tests
